+++ src/components/text-editor/__internal__/text-editor.d.ts
@@ -1,0 +1,14 @@
+import * as React from 'react';
+
+export interface TextEditorProps {
+  characterLimit?: number;
+  labelText: string;
+  onChange: (event: object) => void;
+  onCancel?: () => void;
+  onSave?: () => void;
+  value: object;
+}
+
+declare const TextEditor: React.FunctionComponent<TextEditorProps>;
+
+export default TextEditor;